Finishes cart page for logged in users.
The cart page currently works for logged in users as intended. It does
not work and will give errors if not logged in, as it expects to be
able to get tax rates from the province of the logged in user. Need to
add logic to not display tax rate unless logged in.